Implemented BuyerService

+++ buyer-service/src/main/java/com/iihtproject/buyerservice/dto/BidDto.java
@@ -49,5 +49,5 @@ public class BidDto {
     private String productId;
 
     @NotNull
-    private Long bidAmount;
+    private Double bidAmount;
 }
